fix(catalog): foto na vitrine + estoque decimal entre serviços
## Vitrine sem foto
A listagem de produtos não devolvia imagem nenhuma — só o detalhe devolvia.
O card caía no emoji da categoria, e ninguém compra uma furadeira de R$ 429
olhando "⚒". Foto na vitrine é o que converte.

loadThumbnails preenche a capa de uma página inteira em UMA query
(`= ANY($1)` + `DISTINCT ON`). Chamar loadImages por produto seria N+1 —
com 100 itens por página a vitrine viraria o endpoint mais caro do sistema.
Só a capa entra na listagem; a galeria segue exclusiva do detalhe.
Falha aqui degrada para o ícone, nunca derruba a listagem.

## Bug de contrato entre serviços (quebraria em produção)
A migration 005 trocou products.stock de INT para NUMERIC(14,3) para permitir
venda fracionada (2,5 m de cabo, 1,5 m³ de areia). O order-service decodificava
em `Stock int`, e json.Unmarshal recusa fracionário em inteiro.

O build passava nos dois lados. Só quebraria em runtime, com dado real — e como
o erro vem do decode e não da regra de negócio, sairia como "erro ao criar
pedido" genérico, sem apontar o estoque. Corrigida a cadeia inteira
(Product.Stock, Shortage.Available, comparação) e a mensagem ao cliente agora
formata em pt-BR: "1,5" e não "1.5", "200" e não "200,000".

## Perfil de importação com kind vazio dava 500
`nullIfEmpty(in.Kind)` mandava NULL explícito para uma coluna
NOT NULL DEFAULT 'generic'. No Postgres o DEFAULT só vale quando a coluna é
OMITIDA do INSERT — NULL explícito viola a constraint. Todo perfil criado sem
`kind` (o caso comum: planilha de fornecedor) falhava.

## Infra local
Bancos, Redis e Redpanda presos em 127.0.0.1 — estavam em 0.0.0.0, expostos a
qualquer aparelho da rede local com usuário/senha `utilar`.

## Scripts de ingestão
- importar_curado.py: 285 produtos curados, upsert idempotente por SKU,
  entra como rascunho (publicar é decisão humana, nunca efeito de importação)
- imagens_commons.py: fotos CC0/domínio público do Wikimedia — dado de teste,
  não catálogo final. Nenhuma base pública brasileira tem foto de produto.

## Auditoria arquitetural
docs/security/auditoria-arquitetural-2026-07-18.md: dois achados de desenho —
JWT_SECRET compartilhado permite que qualquer serviço comprometido emita token
de admin, e DEV_MODE=true em produção entrega o sistema via header HTTP.

+++ services/order-service/internal/catalogclient/client.go
@@ -27,7 +27,14 @@ type Product struct {
 	ID    string  `json:"id"`
 	Name  string  `json:"name"`
 	Price float64 `json:"price"`
-	Stock int     `json:"stock"`
+
+	// Stock é float64, não int: a migration 005 do catalog trocou
+	// products.stock para NUMERIC(14,3) pra permitir venda fracionada (2,5 m
+	// de cabo, 1,5 m³ de areia). Decodificar em int faz json.Unmarshal recusar
+	// qualquer produto com saldo fracionado, e o erro sai como falha genérica
+	// de criação de pedido — sem dizer que o problema é o estoque.
+	// Coberto por stockdecimal_test.go.
+	Stock float64 `json:"stock"`
 }
 
 var (
